color-profile-service/requirements.txt
flask==3.0.3
PyMuPDF==1.24.9
pillow==10.4.0
numpy==1.26.4
scikit-learn==1.5.1
pikepdf==9.0.0

